// src/store/user/user.types.ts
enum USER_ACTION_TYPES {
  SET_CURRENT_USER = "user/SET_CURRENT_USER",
  CHECK_USER_SESSION = "user/CHECK_USER_SESSION",
  USER_NOT_FOUND_IN_SESSION = "user/USER_NOT_FOUND_IN_SESSION",
  GOOGLE_SIGN_IN_START = "user/GOOGLE_SIGN_IN_START",
  EMAIL_SIGN_IN_START = "user/EMAIL_SIGN_IN_START",
  SIGN_IN_SUCCESS = "user/SIGN_IN_SUCCESS",
  SIGN_IN_FAIL = "user/SIGN_IN_FAIL",
  SIGN_OUT_START = "user/SIGN_OUT_START",
  SIGN_UP_START = "user/SIGN_UP_START",
}

export default USER_ACTION_TYPES;